improve like php, js, database

+++ src/views/pushLike.php
@@ -4,16 +4,18 @@ include_once('../db/database.php');
 $database = new Database();
 
 // Verifica se sono stati ricevuti i dati corretti tramite POST
-if (isset($_POST['postID']) && isset($_POST['userID']) isset($_POST['add'])) {
+if (isset($_POST['postID']) && isset($_POST['userID']) && isset($_POST['add'])) {
     $postID = $_POST['postID'];
+    $userID = $_POST['userID'];
     $add = $_POST['add'];
     
     if($add){
         if(!$database->insertLike($postID, $userID)){
-            echo "Si è verificato un errore durante l'inserimetno nel database."
-        } else 
+            echo "Si è verificato un errore durante l'inserimetno nel database.";
+        } 
+    } else {
         if(!$database->removeLike($postID, $userID)){
-            echo "Si è verificato un errore durante la rimozione del like dal database."
+            echo "Si è verificato un errore durante la rimozione del like dal database.";
         }
     }
 } else {
